counter: always use long for dimension indexes, int64_t for add value

+++ include/counter/counter.h
@@ -24,16 +24,16 @@ void lttng_counter_destroy(struct lib_counter *counter);
 
 int lttng_counter_read(const struct lib_counter_config *config,
 		      struct lib_counter *counter,
-		      size_t *dimension_indexes,
+		      long *dimension_indexes,
 		      int cpu, int64_t *value,
 		      bool *overflow, bool *underflow);
 int lttng_counter_aggregate(const struct lib_counter_config *config,
 			    struct lib_counter *counter,
-			    size_t *dimension_indexes,
+			    long *dimension_indexes,
 			    int64_t *value,
 			    bool *overflow, bool *underflow);
 int lttng_counter_clear(const struct lib_counter_config *config,
 			struct lib_counter *counter,
-			size_t *dimension_indexes);
+			long *dimension_indexes);
 
 #endif /* _LTTNG_COUNTER_H */
